refactor: put pagemap stuff into its own class
